Restructuring of html and css
I reworked all of the html and css so that the "Holy Grail" layout can
be achieved. There are still things that need to be done, but this is a
great stopping point for now.

I need to figure out how to appropriately size "#pup-quiz" so that the
button images scale without devouring the page.